% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{convert_to_factors}
\alias{convert_to_factors}
\title{Convert multiple columns to factors + Converts specified columns in a data frame to factor type.}
\usage{
convert_to_factors(data, columns)
}
\arguments{
\item{data}{A data frame}

\item{columns}{A character vector of column names to convert to factors}
}
\value{
A data frame with specified columns converted to factors
}
\description{
Convert multiple columns to factors + Converts specified columns in a data frame to factor type.
}
\examples{
df <- data.frame(
  animal_type = c("dog", "cat", "bird"),
  sex = c("M", "F", "M"),
  adopted = c("Yes", "No", "Yes")
)
convert_to_factors(df, c("animal_type", "sex", "adopted"))
}
