% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_data}
\alias{load_data}
\title{Load CSV and print dimensions}
\usage{
load_data(path, verbose = TRUE)
}
\arguments{
\item{path}{A string specifying the path to the CSV file}

\item{verbose}{logical: show information be printed to console (Default TRUE)}
}
\value{
A tibble (data.frame) loaded from the file and print dimensions of the dataset if verbose is TRUE
}
\description{
Load CSV and print dimensions
}
\examples{
data <- load_data("data/processed/longbeach_transformed.csv")
data <- load_data("data/processed/my_data.csv", verbose = FALSE)

}
