% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot confusion matrix as a heatmap
Creates a heatmap visualization of a confusion matrix. Optionally saves the plot to a file.}
\usage{
plot_confusion_matrix(
  cm,
  path_saved = NULL,
  color_low = "#f1f1f1",
  color_high = "#1f77b4",
  text_color = "white",
  text_size = 8,
  title = "Confusion Matrix Heatmap",
  width = NULL,
  height = NULL
)
}
\arguments{
\item{cm}{A caret confusionMatrix object}

\item{path_saved}{File path to save the plot. If NULL, plot is only returned but not saved.}

\item{color_low}{Color for low values in gradient. Default: "#f1f1f1" (light gray)}

\item{color_high}{Color for high values in gradient. Default: "#1f77b4" (blue)}

\item{text_color}{Color for the text labels. Default: "white"}

\item{text_size}{Size for the text labels. Default: 8}

\item{title}{Plot title. Default: "Confusion Matrix Heatmap"}

\item{width}{Plot width in inches when saving. Default: 8}

\item{height}{Plot height in inches when saving. Default: 7}
}
\value{
Invisibly returns the ggplot object.
}
\description{
Plot confusion matrix as a heatmap
Creates a heatmap visualization of a confusion matrix. Optionally saves the plot to a file.
}
\examples{
plot_confusion_matrix(cm)
plot_confusion_matrix(cm, path_saved = "cm_plot.png")

}
